+++ src/tasks/mocha/__tests__/coverage-information-with-failed-test/gulpfile.js
@@ -1,0 +1,15 @@
+'use strict';
+
+const gulp = require('gulp');
+const task = require('../../index');
+
+task.dependencies = {};
+task.register(gulp, 'mocha', {
+  files: [
+    'index-test.js',
+    ['parallel-one-test.js'],
+    ['parallel-two-test.js']
+  ],
+  coverageReporters: ['text-summary'],
+  coverageFrom: 'index.js'
+});