+++ src/tasks/mocha/__tests__/coverage-information-with-failed-test/index-test.js
@@ -1,0 +1,9 @@
+import assert from 'assert';
+import { covered } from './index';
+
+describe('suite', () => {
+  it('caseA', () => {
+    assert.equal(covered(), 0);
+    assert.equal(false, true);
+  });
+});