+++ src/tasks/mocha/__tests__/coverage-information-with-failed-test/index.js
@@ -1,0 +1,11 @@
+'use strict';
+
+export function covered(by) {
+  if (by === 'one') {
+    return 1;
+  } else if (by === 'two') {
+    return 2;
+  }
+
+  return 0;
+}