+++ src/tasks/mocha/__tests__/coverage-information-with-failed-test/parallel-one-test.js
@@ -1,0 +1,8 @@
+import assert from 'assert';
+import { covered } from './index';
+
+describe('suite', () => {
+  it('caseOne', () => {
+    assert.equal(covered('one'), 1);
+  });
+});