+++ src/tasks/mocha/__tests__/runner-init-failure/gulpfile.js
@@ -1,0 +1,10 @@
+'use strict';
+
+const gulp = require('gulp');
+const path = require('path');
+const task = require('../../index');
+
+task.dependencies = {};
+task.register(gulp, 'mocha', {
+  require: path.join(__dirname, 'setup-file.js')
+});