+++ src/tasks/mocha/messages.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  INIT: 'INIT',
+  WORK: 'WORK',
+  DONE: 'DONE',
+  CLOSE: 'CLOSE'
+};